# app/Http/Controllers/API/UetdsCityController.php
<?php

namespace App\Http\Controllers\API;

use App\Http\Controllers\Controller;
use App\Repositories\UetdsCityRepository;
use Illuminate\Http\Request;
use App\Traits\ApiResponser;
use Illuminate\Http\Response;
use Illuminate\Support\Facades\Validator;

class UetdsCityController extends Controller
{
    use ApiResponser;

    private UetdsCityRepository $uetdsCityRepository;

    public function __construct(UetdsCityRepository $uetdsCityRepository)
    {
        $this->uetdsCityRepository = $uetdsCityRepository;
    }

    /**
     * @OA\Get(
     *      path="/api/companies/cities?page={number}",
     *      operationId="getCitiesList",
     *      tags={"UetdsCities"},
     *      security={ {"bearerAuth":{} }},
     *      summary="Get list of Cities",
     *      description="Returns list of Cities",
     *      @OA\Parameter(
     *          name="number",
     *          description="Page Number",
     *          required=false,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *      ),
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="success"),
     *              @OA\Property(property="message", type="string", example=null),
     *              @OA\Property(property="data", type="string", example="array of Cities list"),
     *          )
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="Unauthenticated")
     *          )
     *     )
     * )
     */
    public function index()
    {
        $result = $this->uetdsCityRepository->getAll();
        return $this->successResponse($result);
    }

    /**
     * @OA\Post(
     *      path="/api/companies/cities/create",
     *      operationId="storeCities",
     *      tags={"UetdsCities"},
     *      security={ {"bearerAuth":{} }},
     *      summary="Create New City",
     *      description="Returns City data",
     *      @OA\RequestBody(
     *          required=true,
     *          description="I just fill Required Fields",
     *          @OA\MediaType(
     *              mediaType="multipart/form-data",
     *              @OA\Schema(
     *                  required={"city_code","city","zone_code","zone"},
     *                  @OA\Property(property="city_code", type="number", format="city_code", example="34"),
     *                  @OA\Property(property="city", type="string", format="city", example="Istanbul"),
     *                  @OA\Property(property="zone_code", type="number", format="zone_code", example="55"),
     *                  @OA\Property(property="zone", type="string", format="zone", example="Sisli"),
     *              )
     *          ),
     *      ),
     *      @OA\Response(
     *          response=201,
     *          description="Successful operation",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="success"),
     *              @OA\Property(property="message", type="string", example="City Successfully Created"),
     *              @OA\Property(property="data", type="string", example="array of City list"),
     *          )
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="Unauthenticated")
     *          )
     *     ),
     *      @OA\Response(
     *          response=203,
     *           description="Validation Error response",
     *           @OA\JsonContent(
     *          @OA\Property(property="status", type="string", example="error"),
     *               @OA\Property(property="message", type="string", example="Validation error Message")
     *          )
     *     ),
     *      @OA\Response(
     *          response=422,
     *          description="Unprocessable Content",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="error"),
     *              @OA\Property(property="message", type="string", example="!Something went wrong please try again later."),
     *          )
     *     )
     * )
     */
    public function store(Request $request)
    {
        $data = $request->only("city_code","city","zone_code","zone");
        $validator = Validator::make($data,[
            'city_code' => 'required|integer',
            'city' => 'required|string|max:255',
            'zone_code' => 'nullable|integer',
            'zone' => 'required|string|max:255',
        ]);
        if($validator->fails()){
            return $this->errorResponse($validator->messages(), Response::HTTP_NON_AUTHORITATIVE_INFORMATION);
        }
        $result = $this->uetdsCityRepository->create($data);
        if($result) {
            return $this->successResponse($result, __('response_messages.uetdsCities.created'), Response::HTTP_CREATED);
        }
        return $this->errorResponse(__('response_messages.common.error'), Response::HTTP_UNPROCESSABLE_ENTITY);
    }

    /**
     * @OA\Get(
     *      path="/api/companies/cities/detail/{id}",
     *      operationId="getCityById",
     *      tags={"UetdsCities"},
     *      security={ {"bearerAuth":{} }},
     *      summary="Get of City By Id",
     *      description="Get City Data by City Id",
     *      @OA\Parameter(
     *          name="id",
     *          description="City Id",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *      ),
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="success"),
     *              @OA\Property(property="message", type="string", example=null),
     *              @OA\Property(property="data", type="string", example="array of city"),
     *          )
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="Unauthenticated")
     *          )
     *     ),
     *      @OA\Response(
     *          response=404,
     *          description="Page Not Found",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="error"),
     *              @OA\Property(property="message", type="string", example="Page Not Found"),
     *          )
     *      )
     * )
     */
    public function show($id)
    {
        $result = $this->uetdsCityRepository->getById($id);
        if($result){
            return $this->successResponse($result);
        }
        return $this->errorResponse(__('response_messages.common.404'),Response::HTTP_NOT_FOUND);
    }

    /**
     * @OA\Post(
     *      path="/api/companies/cities/update/{id}",
     *      operationId="updateCity",
     *      tags={"UetdsCities"},
     *      security={ {"bearerAuth":{} }},
     *      summary="Update City",
     *      description="Returns City data",
     *      @OA\Parameter(
     *          name="id",
     *          description="city Id",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *      ),
     *      @OA\RequestBody(
     *          required=true,
     *          description="I just fill Required Fields",
     *          @OA\MediaType(
     *              mediaType="multipart/form-data",
     *              @OA\Schema(
     *                  required={"city_code","city","zone_code","zone"},
     *                  @OA\Property(property="city_code", type="number", format="city_code", example="34"),
     *                  @OA\Property(property="city", type="string", format="city", example="Istanbul"),
     *                  @OA\Property(property="zone_code", type="number", format="zone_code", example="55"),
     *                  @OA\Property(property="zone", type="string", format="zone", example="Sisli"),
     *              )
     *          ),
     *      ),
     *      @OA\Response(
     *          response=201,
     *          description="Successful operation",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="success"),
     *              @OA\Property(property="message", type="string", example="City Successfully Updated"),
     *              @OA\Property(property="data", type="string", example="array of city data"),
     *          )
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="Unauthenticated")
     *          )
     *     ),
     *      @OA\Response(
     *          response=404,
     *          description="Page Not Found",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="error"),
     *              @OA\Property(property="message", type="string", example="Page Not Found"),
     *          )
     *      ),
     *      @OA\Response(
     *          response=203,
     *           description="Validation Error response",
     *           @OA\JsonContent(
     *          @OA\Property(property="status", type="string", example="error"),
     *               @OA\Property(property="message", type="string", example="Validation error Message")
     *          )
     *     ),
     *      @OA\Response(
     *          response=422,
     *          description="Unprocessable Content",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="error"),
     *              @OA\Property(property="message", type="string", example="!Something went wrong please try again later."),
     *          )
     *     )
     * )
     */
    public function update(Request $request, $id)
    {
        $data = $request->only("city_code","city","zone_code","zone");
        $validator = Validator::make($data,[
            'city_code' => 'required|integer',
            'city' => 'required|string|max:255',
            'zone_code' => 'nullable|integer',
            'zone' => 'required|string|max:255',
        ]);
        if($validator->fails()){
            return $this->errorResponse($validator->messages(), Response::HTTP_NON_AUTHORITATIVE_INFORMATION);
        }
        $result = $this->uetdsCityRepository->update($id,$data);
        if($result){
            $result = $this->uetdsCityRepository->getById($id);
            return $this->successResponse($result, __('response_messages.uetdsCities.updated'));
        }
        return $this->errorResponse(__('response_messages.common.404'),Response::HTTP_NOT_FOUND);
    }

    /**
     * @OA\Post(
     *      path="/api/companies/cities/delete/{id}",
     *      operationId="destoryCity",
     *      tags={"UetdsCities"},
     *      security={ {"bearerAuth":{} }},
     *      summary="Remove city",
     *      description="Remove City by Id",
     *      @OA\Parameter(
     *          name="id",
     *          description="City Id",
     *          required=true,
     *          in="path",
     *          @OA\Schema(
     *              type="integer"
     *          )
     *      ),
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="success"),
     *              @OA\Property(property="message", type="string", example="City Deleted Successfully"),
     *              @OA\Property(property="data", type="string", example=null),
     *          )
     *       ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *          @OA\JsonContent(
     *              @OA\Property(property="message", type="string", example="Unauthenticated")
     *          )
     *     ),
     *      @OA\Response(
     *          response=404,
     *          description="Page Not Found",
     *          @OA\JsonContent(
     *              @OA\Property(property="status", type="string", example="error"),
     *              @OA\Property(property="message", type="string", example="Page Not Found"),
     *          )
     *      )
     * )
     */
    public function destroy($id)
    {
        $result = $this->uetdsCityRepository->delete($id);
        if($result)
        {
            return $this->successResponse(null, __('response_messages.uetdsCities.deleted'));
        }
        return $this->errorResponse(__('response_messages.common.404'),Response::HTTP_NOT_FOUND);
    }
}